Adding sweetalert2 functions

+++ server/public/scripts/controllers/events.controller.js
@@ -15,7 +15,15 @@ myApp.controller('EventController', function($location, $http, UserAuthService, 
     console.log('addEvent function clicked', newEvent);
     $http.post('/events/create', newEvent)
     .then(function(response){
+      addEventAlert();
       console.log('added event', response);
+       }).catch(function(err){
+        swal(
+          'Oops...',
+          'Something went wrong!',
+          'error'
+        );
+
     });
   };
 
@@ -32,17 +40,48 @@ myApp.controller('EventController', function($location, $http, UserAuthService, 
   vm.editEvent = function(selectedEvent){
     console.log( 'in editEvents functon', selectedEvent);
     $http.put('/events/edit', selectedEvent).then(function(selectedEvent){
+      editEventAlert();
       getEvents();
-    }); // end success
+
+    }).catch(function(err){
+     swal(
+       'Oops...',
+       'Something went wrong!',
+       'error'
+     );
+
+ }); // end success
   }; // end editEvent
 
   //Delete an event from the Events table
   vm.deleteThisEvent = function(selectedEvent){
     console.log( 'in deleteEvents function', selectedEvent);
     vm.data.selectedEvent = selectedEvent;
+    swal({
+    title: 'Are you sure?',
+    text: "This will delete this event and data associated with it.",
+    type: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#32CD32',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'Yes, delete it!'
+  }).then(function () {
     $http.delete('/events/edit/' + selectedEvent.id).then(function(selectedEvent){
       getEvents();
-    }); // end success
+    });
+    swal(
+        'Deleted!',
+        'Your event has been deleted.',
+        'success'
+      );
+  }).catch(function(err){
+   swal(
+     'Oops...',
+     'Something went wrong!',
+     'error'
+   );
+
+});
   };// end deleteThisEvent
 
 
@@ -53,5 +92,25 @@ myApp.controller('EventController', function($location, $http, UserAuthService, 
     $location.url('/events/checkIn');
   };
 
+// SweetAlert2 Functions
+function addEventAlert() {
+  swal({
+    title: "Success!",
+    text: "This event has been added",
+    confirmButtonText: "View All Events",
+    type: "success"
+  }).then(function() {
+    window.location.href = "#/events";
+  });
+}
+
+function editEventAlert() {
+    swal({
+      title: "Success!",
+      text: "This event has been updated",
+      confirmButtonText: "View Events",
+      type: "success"
+    });
+  }
 
 });
